// backend/src/common/enum/relationship-type.enum.ts
export enum RelationshipType {
  HUSBAND = 'Husband',
  WIFE = 'Wife',
  SPOUSE = 'Spouse',

  FATHER = 'Father',
  MOTHER = 'Mother',
  PARENT = 'Parent',

  SON = 'Son',
  DAUGHTER = 'Daughter',
  CHILD = 'Child',

  BROTHER = 'Brother',
  SISTER = 'Sister',
  SIBLING = 'Sibling',

  BUSINESS_PARTNER = 'Business Partner',
  TEAM_MEMBER = 'Team Member',

  OTHER = 'Other',
}